update k  type to np.float32

+++ opt.py
@@ -10,6 +10,6 @@ def get_opt():
     parser.add_argument('--test_dir',type=str,default=r'data\WFLW_annotations\WFLW_annotations\list_98pt_rect_attr_train_test\list_98pt_rect_attr_test.txt')
     parser.add_argument('--model_save_path',type=str,default=r'./models')
     parser.add_argument('--log_dir',type=str,default=r'./logs')
-    parser.add_argument('--k',type=int,default=1)
+    parser.add_argument('--k',type=np.float32,default=1)
     parser.add_argument('--scheduler_step',type=int,default=10)
     return parser.parse_args()
